Moved validation stuff into mam4/ directory and pruned Fortran stuff.
