+++ .gitmodules
@@ -14,5 +14,5 @@
 	path = ext/skywalker
 	url = git@github.com:eagles-project/skywalker
 [submodule "validation/mam_x_validation"]
-	path = validation/mam_x_validation
+	path = haero/mam4/validation/mam_x_validation
 	url = git@github.com:eagles-project/mam_x_validation
