+++ CMakeLists.txt
@@ -297,14 +297,6 @@ if (HAERO_ENABLE_CHEMISTRY)
   add_subdirectory(chem_driver)
 endif()
 
-# Validation tools (currently offered only for CPU builds with pack size 1,
-# because we're using it with Fortran.)
-#if(NOT HAERO_USE_CUDA AND HAERO_PACK_SIZE EQUAL 1)
-#  # Aerosol cross-validation materials
-#  message(STATUS "Building validation programs...")
-#  add_subdirectory(validation)
-#endif()
-
 # Formatting and format checking using clang-format.
 find_program(CLANG_FORMAT clang-format)
 if (NOT CLANG_FORMAT STREQUAL "CLANG_FORMAT-NOTFOUND")
