+++ haero/mam4/CMakeLists.txt
@@ -1,3 +1,5 @@
 # MAM4 process unit tests.
 add_subdirectory(tests)
 
+# MAM4 validation via Skywalker drivers.
+add_subdirectory(validation)
