+++ haero/mam4/tests/mam4_nucleation_unit_tests.cpp
@@ -1,4 +1,4 @@
-#include <haero/mam4/mam4.hpp>
+#include <haero/mam4.hpp>
 
 #include <catch2/catch.hpp>
